+++ src/views/Home/index.tsx
@@ -1,4 +1,4 @@
-import logo from "../logo.svg";
+import logo from "../../logo.svg";
 import "./style.css";
 
 export function Home() {
